Added subscriptions search type
