+++ src/acf/acf_init.php
@@ -8,7 +8,7 @@ require __DIR__.'/options_page.php';
 //  ┌─────────────────────────────────────────────────────────────────────────┐
 //  │                          The ACF Dashboard                              │
 //  └─────────────────────────────────────────────────────────────────────────┘
-require __DIR__.'/dashboard.php';
+// require __DIR__.'/dashboard.php';
 
 //  ┌─────────────────────────────────────────────────────────────────────────┐
 //  │                      The ACF import/export tab page                     │
